requirements.txt
# Streamlit
streamlit
# Airtable
pyairtable
# GCP
google-cloud-storage
# Data handling
pandas
plotly
# Encoding
bcrypt